Refactor react burger stacker to functional components
Also solved the reverse bug. Props, like state, should not be
mutated in order for React to render properly. Therefore we should
either a) only call reverse on the eventual component array, not
the prop array, so we are not mutating props, or b) create a copy
of the props array which we can directly mutate so as to leave the
original array of props untouched.